requirements.txt
pymongo
langchain
langchain_openai
langchain_community
python-dotenv